Flask App/requirements.txt
google-cloud-storage==2.8.0
flask==2.0.1
joblib==1.2.0
# pyLDAvis==1.5.0
pyLDAvis==3.4.0
# keras==2.5.0
keras==2.12.0
# tensorflow-cpu==2.5.0
tensorflow-cpu==2.12.0
# protobuf==3.18.0
protobuf==3.20.3
numpy==1.22
scipy==1.10.1
pillow==8.2.0
flask_cors
pyreadstat
sentence_transformers